requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.9
pyyaml>=6.0
itsdangerous>=2.1.0
aiofiles>=23.0.0
pyinstaller>=6.0.0
pystray>=0.19.5
pillow>=10.0.0
